+++ app/(panel)/_layout.tsx
@@ -1,6 +1,7 @@
 import MaterialIcons from '@expo/vector-icons/MaterialIcons';
 import { Tabs } from 'expo-router';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import { Platform } from 'react-native';
 
 export default function TabLayout() {
   const insets = useSafeAreaInsets();
@@ -10,19 +11,22 @@ export default function TabLayout() {
       screenOptions={{
         tabBarActiveTintColor: "#C67C4E",
         tabBarShowLabel: false,
+        tabBarHideOnKeyboard: true,
         tabBarStyle: {
-          position: "absolute",
-          bottom: insets.bottom, // ajuste seguro para evitar a barra nativa
+          position: 'absolute',
           left: 20,
           right: 20,
-          height: 50,
-          backgroundColor: "rgba(255, 255, 255, 0.9)", // leve opacidade
+          bottom: insets.bottom + (Platform.OS === 'android' ? 10 : 0), 
+          height: 60,
+          backgroundColor: "rgba(255, 255, 255, 0.95)",
           borderRadius: 30,
           borderTopWidth: 0,
-          elevation: 0, // sombra no Android
-          shadowColor: "#000", // sombra no iOS
+          elevation: 5,
+          shadowColor: "#000", 
+          shadowOffset: { width: 0, height: 5 },
           shadowOpacity: 0.1,
-          paddingBottom: 10,
+          shadowRadius: 10,
+          paddingBottom: Platform.OS === 'android' ? 10 : 15,
           paddingTop: 10,
         },
       }}
@@ -45,15 +49,38 @@ export default function TabLayout() {
           ),
         }}
       />
-      <Tabs.Screen
-        name="profile"
+      
+       <Tabs.Screen
+        name="cart"
         options={{
           headerShown: false,
           tabBarIcon: ({ color }) => (
-            <MaterialIcons size={28} name="supervised-user-circle" color={color} />
+            <MaterialIcons size={28} name="add-shopping-cart" color={color} />
           ),
         }}
       />
+
+        <Tabs.Screen
+        name="orders"
+        options={{
+          headerShown: false,
+          tabBarIcon: ({ color }) => (
+            <MaterialIcons size={28} name="view-list" color={color} />
+          ),
+        }}
+      />
+
+      <Tabs.Screen
+        name="account"
+        options={{
+          headerShown: false,
+          tabBarIcon: ({ color }) => (
+            <MaterialIcons size={28} name="person-outline" color={color} />
+          ),
+        }}
+      />
+
+       
     </Tabs>
   );
 }
