+++ app/(panel)/cart.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+
+const Cart = () => {
+    return (
+        <View>
+            
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({})
+
+export default Cart;
