+++ app/_layout.tsx
@@ -1,9 +1,9 @@
-import { Stack, useRouter, useSegments } from 'expo-router';
-import * as SplashScreen from 'expo-splash-screen';
-import * as Font from 'expo-font';
-import { useEffect, useState } from 'react';
-import { ActivityIndicator, View } from 'react-native';
-import TokenContextProvider, { useTokenContext } from '@/context/useContext';
+import { Stack, useRouter, useSegments } from "expo-router";
+import * as SplashScreen from "expo-splash-screen";
+import * as Font from "expo-font";
+import { useEffect, useState } from "react";
+import { ActivityIndicator, View } from "react-native";
+import TokenContextProvider, { useTokenContext } from "@/context/useContext";
 
 SplashScreen.preventAutoHideAsync();
 
@@ -13,11 +13,10 @@ function RootLayoutInner() {
   const segments = useSegments();
   const router = useRouter();
 
-  // Carregar fonte
   useEffect(() => {
     async function loadFonts() {
       await Font.loadAsync({
-        'Sora': require('@/assets/fonts/Sora-VariableFont_wght.ttf'),
+        Sora: require("@/assets/fonts/Sora-VariableFont_wght.ttf"),
       });
 
       setFontsLoaded(true);
@@ -27,28 +26,26 @@ function RootLayoutInner() {
     loadFonts();
   }, []);
 
-  // Redirecionamento baseado em autenticação
   useEffect(() => {
     if (!fontsLoaded || isLoading) return;
 
-    const inPublicRoute = segments[0] !== '(panel)';
+    const inPublicRoute = segments[0] !== "(panel)";
 
     if (token && inPublicRoute) {
-      router.replace('/(panel)/home');
+      router.replace("/(panel)/home");
     } else if (!token && !inPublicRoute) {
-      router.replace('/');
+      router.replace("/");
     }
   }, [segments, token, isLoading, fontsLoaded]);
 
-  // Mostrar bolinha girando enquanto carrega
   if (isLoading || !fontsLoaded) {
     return (
       <View
         style={{
           flex: 1,
-          justifyContent: 'center',
-          alignItems: 'center',
-          backgroundColor: '#fff',
+          justifyContent: "center",
+          alignItems: "center",
+          backgroundColor: "#fff",
         }}
       >
         <ActivityIndicator size="large" color="#000" />
@@ -59,9 +56,11 @@ function RootLayoutInner() {
   return (
     <Stack
       screenOptions={{
-        headerShown: false,
+        headerShown: true, // default
       }}
-    />
+    >
+      <Stack.Screen name="(panel)" options={{ headerShown: false }} />
+    </Stack>
   );
 }
